backend/requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
django==5.0.4; python_version >= '3.10'
djangorestframework==3.15.1; python_version >= '3.6'
psycopg2==2.9.9; python_version >= '3.7'
sqlparse==0.5.0; python_version >= '3.8'
tzdata==2024.1; sys_platform == 'win32'
celery
redis